backend: add assertCallReturnValue to mocked hooks in tests
