+++ backend/pkg/api/api_integration_test.go
@@ -16,14 +16,18 @@ import (
 	"context"
 	"encoding/json"
 	"io"
+	"math"
 	"math/rand"
 	"net/http"
+	"runtime"
 	"strconv"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/cloudhut/common/logging"
 	"github.com/cloudhut/common/rest"
+	"github.com/go-chi/chi/v5"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -33,7 +37,11 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/redpanda-data/console/backend/pkg/config"
+	"github.com/redpanda-data/console/backend/pkg/connect"
+	"github.com/redpanda-data/console/backend/pkg/console"
 	"github.com/redpanda-data/console/backend/pkg/testutil"
+
+	rp "github.com/redpanda-data/console/backend/pkg/redpanda"
 )
 
 type APIIntegrationTestSuite struct {
@@ -153,4 +161,368 @@ func (s *APIIntegrationTestSuite) apiRequest(ctx context.Context,
 	assert.NoError(t, err)
 
 	return res, body
+}
+
+type assertCallReturnValue struct {
+	Err *rest.Error
+
+	BoolValue bool
+
+	SliceValue []string
+}
+
+func assertHookCall(t *testing.T) {
+	pc, _, _, _ := runtime.Caller(1)
+	fnName := runtime.FuncForPC(pc).Name()
+	assert.Fail(t, "unexpected call to hook function:"+fnName)
+}
+
+// assertHooks is the default hook for tests that deny everything by default and assert when functions are called
+type assertHooks struct {
+	t *testing.T
+
+	allowedCalls map[string]map[string]bool
+	returnValues map[string]map[string]assertCallReturnValue
+}
+
+func (a *assertHooks) isCallAllowed(topicName string) bool {
+	pc, _, _, _ := runtime.Caller(1)
+	fnName := runtime.FuncForPC(pc).Name()
+	parts := strings.Split(fnName, ".")
+	fnName = parts[len(parts)-1]
+	topicMap, ok := a.allowedCalls[fnName]
+	if !ok || len(topicMap) == 0 {
+		return false
+	}
+
+	if v, ok := topicMap["any"]; ok {
+		return v
+	}
+	return topicMap[topicName]
+
+}
+
+func (a *assertHooks) getCallReturnValue(topicName string) assertCallReturnValue {
+	pc, _, _, _ := runtime.Caller(1)
+	fnName := runtime.FuncForPC(pc).Name()
+	parts := strings.Split(fnName, ".")
+	fnName = parts[len(parts)-1]
+	topicMap, ok := a.returnValues[fnName]
+	if !ok || len(topicMap) == 0 {
+		return assertCallReturnValue{}
+	}
+	if v, ok := topicMap["any"]; ok {
+		return v
+	}
+	return topicMap[topicName]
+}
+
+func newAssertHooks(t *testing.T, returnValues map[string]map[string]assertCallReturnValue) *Hooks {
+	h := &assertHooks{
+		allowedCalls: map[string]map[string]bool{},
+		returnValues: map[string]map[string]assertCallReturnValue{},
+	}
+
+	for n, v := range returnValues {
+		for tn, _ := range v {
+			if len(h.allowedCalls[n]) == 0 {
+				h.allowedCalls[n] = map[string]bool{}
+			}
+
+			h.allowedCalls[n][tn] = true
+		}
+
+		h.returnValues[n] = v
+	}
+
+	return &Hooks{
+		Authorization: h,
+		Route:         h,
+		Console:       h,
+	}
+}
+
+// Router Hooks
+func (a *assertHooks) ConfigAPIRouter(_ chi.Router)      {}
+func (a *assertHooks) ConfigWsRouter(_ chi.Router)       {}
+func (a *assertHooks) ConfigInternalRouter(_ chi.Router) {}
+func (a *assertHooks) ConfigRouter(_ chi.Router)         {}
+
+// Authorization Hooks
+func (a *assertHooks) CanSeeTopic(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+
+	rv := a.getCallReturnValue(topic)
+
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanCreateTopic(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanEditTopicConfig(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteTopic(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanPublishTopicRecords(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteTopicRecords(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanViewTopicPartitions(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanViewTopicConfig(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanViewTopicMessages(_ context.Context, r *ListMessagesRequest) (bool, *rest.Error) {
+	if !a.isCallAllowed(r.TopicName) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(r.TopicName)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanUseMessageSearchFilters(_ context.Context, r *ListMessagesRequest) (bool, *rest.Error) {
+	if !a.isCallAllowed(r.TopicName) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(r.TopicName)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanViewTopicConsumers(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) AllowedTopicActions(_ context.Context, topic string) ([]string, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.SliceValue, rv.Err
+}
+
+func (a *assertHooks) PrintListMessagesAuditLog(_ *http.Request, r *console.ListMessageRequest) {
+	if !a.isCallAllowed(r.TopicName) {
+		assertHookCall(a.t)
+	}
+}
+
+func (a *assertHooks) CanListACLs(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanCreateACL(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteACL(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanListQuotas(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanSeeConsumerGroup(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanEditConsumerGroup(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteConsumerGroup(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) AllowedConsumerGroupActions(_ context.Context, _ string) ([]string, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.SliceValue, rv.Err
+}
+
+func (a *assertHooks) CanPatchPartitionReassignments(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanPatchConfigs(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanViewConnectCluster(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanEditConnectCluster(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteConnectCluster(_ context.Context, topic string) (bool, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) AllowedConnectClusterActions(_ context.Context, topic string) ([]string, *rest.Error) {
+	if !a.isCallAllowed(topic) {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue(topic)
+	return rv.SliceValue, rv.Err
+}
+
+func (a *assertHooks) CanListKafkaUsers(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanCreateKafkaUsers(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) CanDeleteKafkaUsers(_ context.Context) (bool, *rest.Error) {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue, rv.Err
+}
+
+func (a *assertHooks) IsProtectedKafkaUser(_ string) bool {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.BoolValue
+}
+
+// Console hooks
+func (a *assertHooks) ConsoleLicenseInformation(_ context.Context) rp.License {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	return rp.License{Source: rp.LicenseSourceConsole, Type: rp.LicenseTypeOpenSource, ExpiresAt: math.MaxInt32}
+}
+
+func (a *assertHooks) EnabledFeatures() []string {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	rv := a.getCallReturnValue("any")
+	return rv.SliceValue
+}
+
+func (a *assertHooks) EndpointCompatibility() []console.EndpointCompatibilityEndpoint {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	return nil
+}
+
+func (a *assertHooks) EnabledConnectClusterFeatures(_ context.Context, _ string) []connect.ClusterFeature {
+	if !a.isCallAllowed("any") {
+		assertHookCall(a.t)
+	}
+	return nil
 }
